Responsive Design Challenge 1:
The challenge was about preventing overflow and adjusting the
width of the child container. To complete this and fix the issue
the following was done:
1. The height was removed from the parent element
because it limits the flexibility of the child. Additionally,
the margin on the parent was already set to "0, auto".
2. The width on the child container was added using a
percentage of 60% and a padding on "0.5em".

All of these adjustment made the content more flexible
and removed the overflow that was occuring.